Crate jsonrpsee_ws_client

jsonrpsee-ws-client

jsonrpsee-ws-client is a JSON RPC WebSocket client library that’s is built for async/await.

Runtime support

This library uses tokio as the runtime and does not support other kinds of runtimes. Tokio versions v1 and v0.2 are supported behind tokioV1 and tokioV02 feature flags correspondingly.

Re-exports

pub use client::WsClient;

pub use client::WsClientBuilder;

Modules

client
WebSocket Client.

error
Shared error type.

helpers
Helpers.

manager
Request manager. Handles and monitors JSONRPC v2 method calls and subscriptions

stream
Stream. Convenience wrapper for a stream (AsyncRead + AsyncWrite) which can either be plain TCP or TLS.

traits
Traits

transport
WebSocket transport.

v2
JSON-RPC 2.0 specification related types v2.

Structs

BatchMessage
Batch request message.

JsonRawValue
Reference to a range of bytes encompassing a single valid JSON value in the input data.

RegisterNotificationMessage
RegisterNotification message.

RequestMessage
Request message.

Subscription
Active subscription on the client.

SubscriptionMessage
Subscription message.

Enums

Error
Error type.

FrontToBack
Message that the Client can send to the background task.

JsonValue
Represents any valid JSON value.

SubscriptionKind
Subscription kind

Constants

TEN_MB_SIZE_BYTES
Ten megabytes.

Traits

DeserializeOwned
A data structure that can be deserialized without borrowing any data from the deserializer.

Serialize
A data structure that can be serialized into any data format supported by Serde.

Functions

to_json_value
Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

Type Definitions

Cow
Compact three word Cow that puts the ownership tag in capacity. This is a type alias, for documentation see beef::generic::Cow.

Derive Macros

Serialize