Function jsonrpsee_ws_client::helpers::stop_subscription

pub async fn stop_subscription(
    sender: &mut WsSender, 
    manager: &mut RequestManager, 
    unsub: RequestMessage
)

Sends an unsubscribe to request to server to indicate that the client is not interested in the subscription anymore.