Struct jsonrpsee_ws_client::manager::RequestManager

pub struct RequestManager { /* fields omitted */ }

Manages and monitors JSONRPC v2 method calls and subscriptions.

Implementations

impl RequestManager

pub fn new() -> Self

Create a new RequestManager.

pub fn insert_pending_call(
&mut self,
id: u64,
send_back: Option<Sender<Result<JsonValue, Error>>>
) -> Result<(), Option<Sender<Result<JsonValue, Error>>>>

Tries to insert a new pending call.

Returns Ok if the pending request was successfully inserted otherwise Err.

pub fn insert_pending_batch(
&mut self,
batch: Vec<u64>,
send_back: Sender<Result<Vec<JsonValue>, Error>>
) -> Result<(), Sender<Result<Vec<JsonValue>, Error>>>

Tries to insert a new batch request

Returns Ok if the pending request was successfully inserted otherwise Err.

pub fn insert_pending_subscription(
&mut self,
sub_req_id: u64,
unsub_req_id: u64,
send_back: Sender<Result<(Receiver<JsonValue>, SubscriptionId), Error>>,
unsubscribe_method: String
) -> Result<(), Sender<Result<(Receiver<JsonValue>, SubscriptionId), Error>>>

Tries to insert a new pending subscription and reserves a slot for a “potential” unsubscription request.

Returns Ok if the pending request was successfully inserted otherwise Err.

pub fn insert_subscription(
&mut self,
sub_req_id: u64,
unsub_req_id: u64,
subscription_id: SubscriptionId,
send_back: Sender<JsonValue>,
unsubscribe_method: String
) -> Result<(), Sender<JsonValue>>

Tries to insert a new subscription.

Returns Ok if the pending request was successfully inserted otherwise Err.

pub fn insert_notification_handler(
&mut self,
method: &str,
send_back: Sender<JsonValue>
) -> Result<(), Error>

Inserts a handler for incoming notifications

pub fn remove_notification_handler(
&mut self,
method: String
) -> Result<(), Error>

Removes a notification handler

pub fn complete_pending_subscription(
&mut self,
request_id: u64
) -> Option<(u64, Sender<Result<(Receiver<JsonValue>, SubscriptionId), Error>>, String)>

Tries to complete a pending subscription.

Returns Some if the subscription was completed otherwise None.

pub fn complete_pending_batch(&mut self, batch: Vec<u64>) -> Option<BatchState>

Tries to complete a pending batch request

Returns Some if the subscription was completed otherwise None.

pub fn complete_pending_call(
&mut self,
request_id: u64
) -> Option<Option<Sender<Result<JsonValue, Error>>>>

Tries to complete a pending call..

Returns Some if the call was completed otherwise None.

pub fn remove_subscription(
&mut self,
request_id: u64,
subscription_id: SubscriptionId
) -> Option<(u64, Sender<JsonValue>, String, SubscriptionId)>

Tries to remove a subscription.

Returns Some if the subscription was removed otherwise None.

pub fn request_status(&mut self, id: &u64) -> RequestStatus

Returns the status of a request ID

pub fn as_subscription_mut(
&mut self,
request_id: &u64
) -> Option<&mut Sender<JsonValue>>

Get a mutable reference to underlying Sink in order to send messages to the subscription.

Returns Some if the request_id was registered as a subscription otherwise None.

pub fn as_notification_handler_mut(
&mut self,
method: String
) -> Option<&mut Sender<JsonValue>>

Get a mutable reference to underlying Sink in order to send incmoing notifications to the subscription.

Returns Some if the method was registered as a NotificationHandler otherwise None.

pub fn get_request_id_by_subscription_id(
&self,
sub_id: &SubscriptionId
) -> Option<u64>

Reverse lookup to get the request ID for a subscription ID.

Returns Some if the subscription ID was registered as a subscription otherwise None.