Trait jsonrpsee_ws_client::traits::Client

pub trait Client {
#[must_use]    pub fn notification<'a, 'life0, 'async_trait>(
        &'life0 self, 
        method: &'a str, 
        params: JsonRpcParams<'a>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn request<'a, 'life0, 'async_trait, R>(
        &'life0 self, 
        method: &'a str, 
        params: JsonRpcParams<'a>
    ) -> Pin<Box<dyn Future<Output = Result<R, Error>> + 'async_trait + Send, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        R: DeserializeOwned + 'async_trait;

#[must_use]    pub fn batch_request<'a, 'life0, 'async_trait, R>(
        &'life0 self, 
        batch: Vec<(&'a str, JsonRpcParams<'a>), Global>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<R, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        R: DeserializeOwned + Default + Clone + 'async_trait;
}

JSON-RPC client interface that can make requests and notifications.

Required methods

#[must_use]pub fn notification<'a, 'life0, 'async_trait>(
&'life0 self,
method: &'a str,
params: JsonRpcParams<'a>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Send a notification request

#[must_use]pub fn request<'a, 'life0, 'async_trait, R>(
&'life0 self,
method: &'a str,
params: JsonRpcParams<'a>
) -> Pin<Box<dyn Future<Output = Result<R, Error>> + 'async_trait + Send, Global>> where
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
R: DeserializeOwned + 'async_trait,

Send a method call request.

#[must_use]pub fn batch_request<'a, 'life0, 'async_trait, R>(
&'life0 self,
batch: Vec<(&'a str, JsonRpcParams<'a>), Global>
) -> Pin<Box<dyn Future<Output = Result<Vec<R, Global>, Error>> + 'async_trait + Send, Global>> where
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
R: DeserializeOwned + Default + Clone + 'async_trait,

Send a batch request.

The response to batch are returned in the same order as it was inserted in the batch.

Returns Ok if all requests in the batch were answered successfully. Returns Error if any of the requests in batch fails.

Loading content...