Module jsonrpsee_ws_client::v2::params

JSON_RPC params related types.

Structs

JsonRpcNotificationParams
JSON-RPC parameter values for subscriptions.

JsonRpcNotificationParamsAlloc
JSON-RPC parameter values for subscriptions with support for number and strings.

OwnedRpcParams
Owned version of RpcParams.

RpcParams
Parameters sent with the RPC request

TwoPointZero
JSON-RPC v2 marker type.

Enums

Id
Request Id

JsonRpcParams
Serializable JSON-RPC parameters

OwnedId
Owned version of Id that allocates memory for the Str variant.

SubscriptionId
Id of a subscription, communicated by the server.