Crate jsonrpsee_ws_client

jsonrpsee-ws-client

jsonrpsee-ws-client is a JSON RPC WebSocket client library that’s is built for async/await.

Async runtime support

This library uses tokio as the runtime and does not support other runtimes.

Re-exports

pub use jsonrpsee_types as types;

Structs

HeaderMap

A set of HTTP headers

HeaderValue

Represents an HTTP header field value.

PingConfig

Configuration for WebSocket ping/pong mechanism and it may be used to disconnect an inactive connection.

WsClient

Generic asynchronous client.

WsClientBuilder

Builder for WsClient.